[compose] Remove layoutParams from ViewAnnotation composable function. (#2393)
The `ViewGroup.LayoutParams` is a non-compose structure and shouldn't be used as part of the compose API,
the internal `ComposeView` wrapping the `ViewAnnotation.content` should always use `WRAP_CONTENT` size.

In case of tests where the assertion happens before the measurement of the layout, user can force the
content size using `ViewAnnotationOptions.width/height` APIs.